requirements.txt
python-telegram-bot==13.4.1
urllib==1.26.7
